[BACK] Implementing Event to append to Publication
