+++ app/templates/event/show.html.twig
@@ -16,7 +16,7 @@
             </tr>
             <tr>
                 <th>Titre</th>
-                <td>{{ event.publication.title }}</td>
+                <td>{{ event.title }}</td>
             </tr>
             <tr>
                 <th>Description</th>
